Forcefully apply buffer size before starting AU graph
After lots of testing, this appears to force iOS to be compliant to
what we actually want the buffer size to be. I haven't seen a case
yet where iOS gives us a buffer size that is not what we ask for.

I tried a number of other things but this change appears to be
the most reliable way to get iOS to give us what we want.