Do not start the AUGraph after mic permission granted
This fixes a bug on iOS where the AU graph would be started
immediately after the microphone permission was granted.

onMediaServerReset is called immedately after the mic permission
is granted, and was always calling [start] which would start
the AUGraph. Because of this, the iOS codebase implemented a
workaround which was error prone, and it currently causing the
app to freeze after onboarding and the mic permission is granted.

When the app first launches, the augraph isn't started yet, so
when the microphone permission is requested, it is not running,
and does not need to start running.