Try to fix 1104 buffer sizes on iOS
If you have
* iOS 18
* no headset
* Vocal Shortcuts enabled
* Sound Recognition enabled

then iOS returns 1104 frames of audio in the audio processing callback.

This was preventing _any_ audio processing because maximumNumberOfFrames
was capped to 1024.

Superpowered already sets up the buffers to be MAXFRAMES, so we should
be good in terms of buffer size there.

Vocoder can handle all sorts of varying buffer sizes, and iOS does reinit
Vocoder with the iOS buffer size after 20 attempts.